refactor: this module should only convert files from medra to datacite
the upload part has been removed. This module is only converting and
the upload is done with the datacite module.
